Fix module import issue in install.py
This commit addresses an issue in `install.py` where the `comfy` module
could not be found, leading to a `ModuleNotFoundError` when running the
script on some systems, such as Linux.

The issue was caused by the `comfy` directory not being included in the
Python module search path (`sys.path`). To fix this, the parent directory
of `comfy` is added to `sys.path` in `install.py`.

With these changes, the script should work correctly on different
environments, including both Windows and Linux systems.